+++ OpFloat.class.cpp
@@ -2,7 +2,7 @@
 
 OpFloat::OpFloat(std::string value) {
     this->value = value;
-    this->type = Int8;
+    this->type = Float;
 }
 OpFloat::~OpFloat() {
 }
